Revert "Move the logic out of the data context and into the suite validation results"
This reverts commit cb06e72413266676d3d15daaa607ea27edc371f1.
